testbed/simnet: split main.go into per-concern files; tidy adapter comments
main.go was ~920 lines covering flag parsing, dispatch, three workloads,
search, coverage, reporting, and helpers. Split into:

  main.go              — flag parsing + workload dispatch
  adapter.go           — simUDPConn (net.UDPAddr ↔ netip.AddrPort)
  nodes.go             — nodeRec, makeTopic, pickLegacySet, spawnNodes
  coverage.go          — registrationCoverage + multiTopicCoverage helpers
  search.go            — searchResult, runSearches, runMultiTopicSearches
  report.go            — report, writeMetrics, reportMultiTopic
  util.go              — percentile, fatalf, minOrZero/medOrZero/maxOrZero
  workload_single.go   — runSingleTopicWorkload
  workload_multi.go    — runMultiTopicWorkload + runRegistrationProbe
  workload_discng.go   — runDiscNGValidationWorkload

Also drop the inline UDPConn-interface quote from adapter.go — it's a
straight quote of p2p/discover/common.go and adds nothing.
